# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ES7-WebServices",
        "version": "1.0.1",
        "title": "ES7 WebService"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1.0/Contenido": {
            "get": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Contenido de una Escuela Sabatica",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "es7",
                    "in": "query",
                    "description": "Escuela Sabatica",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "tipo",
                    "in": "query",
                    "description": "Tipo de Contenido",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
...
        "/v1.0/LeccionDiaria": {
            "get": {
                "tags": ["Lecciones Diarias"],
                "summary": "Lecciones Diarias de una Escuela Sabatica",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "contenido",
                    "in": "query",
                    "description": " Contenido de Escuela Sabatica",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/LeccionDiaria/{id}/": {
            "get": {
                "tags": ["Lecciones Diarias"],
                "summary": "Encontrar un leccion por su Identificador",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Identificador",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "404": {
                        "description": "Leccion no encontrado"
                    }
                }
            }
        },
        "/v1.0/LeccionDiaria/create": {
            "post": {
                "tags": ["Lecciones Diarias"],
                "summary": "Crear una Leccion",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Leccion Diaria",
                    "in": "body",
                    "description": "Leccion Diaria a crear",
                    "required": true,
                    "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "idLesson": {
                                    "type": "integer"
                                },
                                "idDay": {
                                    "type": "integer"
                                },
                                "dayTitle": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "lessonText": {
                                    "type": "string"
                                }
                            }
                        
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/LeccionDiaria/inactive": {
            "patch": {
                "tags": ["Lecciones Diarias"],
                "summary": "Deshabilitar una leccion",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Leccion Diaria",
                    "in": "body",
                    "description": "Leccion Diaria a Deshabilitar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/LeccionDiaria/update": {
            "put": {
                "tags": ["Lecciones Diarias"],
                "summary": "Editar una leccion",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Leccion Diaria",
                    "in": "body",
                    "description": "Leccion Diaria a editar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "idLesson": {
                                "type": "integer"
                            },
                            "idDay": {
                                "type": "integer"
                            },
                            "dayTitle": {
                                "type": "string"
                            },
                            "date": {
                                "type": "string"
                            },
                            "lessonText": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        }
    }
}